+++ docs/components.md
@@ -17,7 +17,7 @@ In this example we drive a cylinder by a certain amount of millimeters.
 
 [dependencies]
 # Include the library configured for the raspberry pi
-stepper_lib = { version = "0.11", features = [ "rasp" ] } 
+stepper_lib = { version = "0.11.4", features = [ "rasp" ] } 
 
 # ...
 ```
@@ -96,7 +96,7 @@ Click to show Cargo.toml
 
 [dependencies]
 # Include the library configured for the raspberry pi
-stepper_lib = { version = "0.11", features = [ "rasp" ] } 
+stepper_lib = { version = "0.11.4", features = [ "rasp" ] } 
 
 # ...
 ```
@@ -104,12 +104,8 @@ stepper_lib = { version = "0.11", features = [ "rasp" ] }
 <p></p>
 
 ```rust
-// Include components and data
-use stepper_lib::{StepperCtrl, StepperConst, SyncComp, Setup};
-use stepper_lib::data::LinkedData;
-use stepper_lib::meas::SimpleMeas;
-// Include the unit system
-use stepper_lib::units::*;
+// Include library
+use stepper_lib::prelude::*;
 
 // Pin declerations (BCM on raspberry pi)
 const PIN_DIR : u8 = 27;
@@ -117,13 +113,13 @@ const PIN_STEP : u8 = 19;
 
 // Define distance and max speed
 const DELTA : Delta = Delta(10.0);      
-const OMEGA : Omega = Omega(15.0);      
+const OMEGA : Omega = Omega(20.0);      
 
 // Defining component structure
 #[derive(Debug)]
 struct MyComp {
-    ctrl : StepperCtrl,
-    ratio : f32
+    ctrl : StepperCtrl,     // The stepper motor built into the component
+    ratio : f32             // The gear ratio, e.g. a spingle or
 }
 
 impl MyComp {
@@ -140,16 +136,17 @@ impl SimpleMeas for MyComp {
 
 impl Setup for MyComp {
     fn setup(&mut self) -> Result<(), stepper_lib::Error> {
-        Ok(())
+        self.ctrl.setup()?;  // Setting up the super component
+        Ok(())      
     }
 }
 
 impl SyncComp for MyComp {
     // Required memebers
-        fn consts<'a>(&'a self) -> &'a stepper_lib::StepperConst {
+        fn consts<'a>(&'a self) -> &'a StepperConst {
             todo!()
         }
-
+        
         fn vars<'a>(&'a self) -> &'a stepper_lib::data::CompVars {
             todo!()     // Not required in this example
         }
@@ -157,15 +154,10 @@ impl SyncComp for MyComp {
         fn link<'a>(&'a self) -> &'a LinkedData {
             todo!()     // Not required in this example
         }
-
-        fn to_json(&self) -> Result<serde_json::Value, serde_json::Error> {
-            todo!()     // Not required in this example
-        }
     //
     
     // Super component (motor)
         // The following two overrides give the library access to the stepper motor controller stored in our component
-        
         fn super_comp(&self) -> Option<&dyn SyncComp> {
             Some(&self.ctrl)
         }
@@ -177,7 +169,6 @@ impl SyncComp for MyComp {
 
     // Ratio
         // The following two overrides cause the library to translate the distance by the ratio we defined for our component
-
         fn gamma_for_super(&self, this_gamma : Gamma) -> Gamma {
             this_gamma * self.ratio     // Distance is translated by the ratio
         }
@@ -187,8 +178,7 @@ impl SyncComp for MyComp {
         }
     // 
 
-    fn drive_rel(&mut self, delta : Delta, speed_f : f32) 
-            -> Result<Delta, stepper_lib::Error> {
+    fn drive_rel(&mut self, delta : Delta, speed_f : f32) -> Result<Delta, stepper_lib::Error> {
         println!("Now driving!"); // Our custom message
 
         let delta_real = self.ctrl.drive_rel(
@@ -203,11 +193,7 @@ impl SyncComp for MyComp {
 fn main() -> Result<(), stepper_lib::Error> {
     // Create the controls for a stepper motor
     let mut comp = MyComp::new(
-        StepperCtrl::new(
-            StepperConst::MOT_17HE15_1504S, 
-            PIN_DIR, 
-            PIN_STEP
-        ),
+        StepperCtrl::new(StepperConst::MOT_17HE15_1504S, PIN_DIR, PIN_STEP),
         2.0 // Example ratio
     );
     // Link the component to a system
@@ -216,12 +202,13 @@ fn main() -> Result<(), stepper_lib::Error> {
         s_f: 1.5    // System safety factor, should be at least 1.0
     }); 
 
-    comp.setup();
+    comp.setup()?;
 
     // Apply some loads
     comp.apply_inertia(Inertia(0.2));
     comp.apply_force(Force(0.10));
-
+    
+    // Limit the component velocity
     comp.set_omega_max(OMEGA);
 
     println!("Staring to move ... ");
@@ -239,4 +226,4 @@ fn main() -> Result<(), stepper_lib::Error> {
 // "
 ```
 
-(Source: "examples/custom_component.rs")
+(See [the rust example]("../examples/custom_component.rs"))
