+++ examples/cylinder.rs
@@ -16,12 +16,8 @@ stepper_lib = { version = \"0.11\", features = [ \"rasp\" ] }
 ```
 "]
 
-// Include components and data
-use stepper_lib::{StepperCtrl, StepperConst, SyncComp, Setup};
-use stepper_lib::comp::Cylinder;
-use stepper_lib::data::LinkedData;
-// Include the unit system
-use stepper_lib::units::*;
+// Include the library
+use stepper_lib::prelude::*;
 
 // Pin declerations (BCM on raspberry pi)
 const PIN_DIR : u8 = 27;
