+++ examples/tongs.rs
@@ -20,20 +20,15 @@ use core::time::Duration;
 
 use std::thread::sleep;
 
-use stepper_lib::{SimpleTool, Tool};
-use stepper_lib::comp::tool::Tongs;
-use stepper_lib::ctrl::servo::ServoDriver;
-use stepper_lib::data::servo::ServoConst;
-
-// Include the unit system
-use stepper_lib::units::*;
+// Include the library
+use stepper_lib::prelude::*;
 
 // Pin declerations (BCM on raspberry pi)
 const PIN_PWM : u8 = 27;
 
 fn main() -> Result<(), stepper_lib::Error> {
     // Create the tongs controller
-    let mut tongs = Tongs::new(
+    let mut tongs = tool::Tongs::new(
         ServoDriver::new(ServoConst::MG996R, PIN_PWM),    // The tongs use a MG996R servo connected to the BCM pin 27   
         0.2,        // when the pwm signal is at 20%, the tongs are open
         0.8,        // when the pwm signal is at 80%, the tongs are closed
