--- src/err.rs
@@ -1,1 +1,0 @@
-// TODO: Add proper errors to library
