+++ src/prelude.rs
@@ -1,0 +1,21 @@
+// Simple imports of the library
+pub use crate::{AsyncComp, Direction, Setup, StepperCtrl, SimpleTool, SyncComp, SyncCompGroup};
+
+pub use crate::comp::{Cylinder, CylinderTriangle, GearJoint, Tool};
+pub use crate::comp::tool;
+
+pub use crate::ctrl::DcMotor;
+pub use crate::ctrl::pwm::PWMOutput;
+pub use crate::ctrl::pin::*;
+pub use crate::ctrl::servo::ServoDriver;
+
+pub use crate::data::{CompVars, LinkedData, StepperConst};
+pub use crate::data::servo::ServoConst;
+
+pub use crate::math::{CurveBuilder, PathBuilder};
+pub use crate::math::force;
+pub use crate::math::inertia;
+
+pub use crate::meas::SimpleMeas;
+
+pub use crate::units::*;
